- Changed the details element to a div for the sake of accesibility, as the details and summary elements can cause many issues. For one, they strip off the semantics of nested elements, and actions like editing tasks caused the the details element to open and close when pressing space. Having to implement workarounds for these is inefficient, and creating an accordion from scratch turned out to be easier than expected.
- Gave the body a width and height of 100% to align everything to the
center.

- Improved the styling of the category form and allowed the user to close
the form in case they don't want to add a new category.

- Created a new and sematic progressbar that allows for transitions to work
when completing a task. It works cross browser now because it is made out of
divs.

- Added an SVG of waves that fills up a glass ball as the user completes
tasks. It is quite minimalist, and not what I want to use as the main
illustration, but it works properly for now to showcase the concept.