Added archive functionality.
A button appears after the user has completed a task, prompting them to
archive it. They can enter a title and notes, as well as the tasks they
have completed on a given date.
    After a task has been archived, the styling changes and checkbox
states can't get altered anymore. The progress state remains the same,
until the user decides to delete the task, which can get accessed through
the archive.

Upcoming tasks:
- Add a navigation bar to the side
- Create the archive page
- Use the VueRouter

P.S. A clickOutside directive was created to close the archive entry
modal when needed.