Vue Router Implementation
With the help of the router the user can now switch between the tasks and
archive views. A lot of code had to be moved into a seperate component for
the tasks view to be accessible through a .view file.
    The next task will be to figure out the different sorting options for
the archive, and creating appropriate views. Icons for the nav bar will
get added later on as well.