Began styling the archive.
The calendar has been placed to the left side of each entry, and has a
sticky position to allow for it to follow the associated entry. Some
basic styling has been applied to the contents of the entries, but more
work will be done on them once I implement the markdown parser for the
user to write more dynamic content.